Added parts, created board, placed parts and rails
Added Spark core, 5V regulator, N-channel MOSFET for switching the 24V
line for the garage door, and some screw terminals that I can connect to
the Magnetic switch
(http://www.amazon.com/Directed-Electronics-8601-Magnetic-Switch/dp/B0009SUF08/ref=sr_1_2?ie=UTF8&qid=1397978783&sr=8-2&keywords=reed+switch),
a range finder (https://www.sparkfun.com/products/8503), and an RGB LED
(https://www.sparkfun.com/products/11820).  These parts will allow me to
sense the position of the garage door, open and close the garage door,
and detect distance from the wall that the car is and shine a green,
yellow, or red light to show whether to pull forward, slow down, or stop
so that the car is far enough in the garage but not too close to the
wall.  There are 2 connections for a range finder and LED so that a 2
cars can be parked but only 1 set of door open/close and sensors (a
typical 2 car garage setup).