Moved Mosfet and shrunk design
Moved the MOSFET underneath the Spark on the bottom layer so I can save
some space.  I moved just about everything around and was able to shrink
the board down to basically just the size of the Spark and the extra
width of the screw terminals.  It is very small now and costs almost $4
less to build.