Removed MOSFET inline resistors
The garage door button needs to have a direct connection from power
(btn_h) to ground (btn_l) in order to open or close the garage door.
The high side resistor kept this from happening.  I am not sure what I
was thinking with the low side 'pull down'.  This was just a weird
resistor connecting a groundplane on the opener to a ground plane on my
board.

I also renamed a few of the connection labels on the backside to be
correct.  The btn_p and btn_n were backwards and this was causing the
opener to not work.  The range finders also need to have different
labels so that the user knows which one is connected for configuring the
app/Spark Core.